Attempt to fix frame distance mapping.
After help from Klaudiusz <info@rlv-stormarn.de>, the guy responsible for RLVWorkbench (http://rlv-stormarn.de/rlv-workbench/), I managed to understand how the mapping works.

The mapping seems to be off still for some videos, most notably for IT_GiroMortirolo. The start for the Gavia is about 200 meters short. I'll need to find the reason for that.