Depend on libusb-1.0, but use libusb-0.1 API, using libusb-compat
Because of problems using libusb-0.1, usb_bulk_read was sometimes blocking instead
of timing out, I now added the code of libusb-compat to the source and
depend on that.

Using this, we can, on Linux, use the libusb-0.1 API, but with the libusb-1.0
library. This seems to work very nicely now.