Don't send "unassign channel" twice if ANT+ stick closes the channel.
ANT USB1 sticks (serial) do not close the channel by themselves, but USB2
sticks do. Both should work without sending messages in the wrong state.

After a channel has a search timeout, we simply unassign it. If the ANT+
usb stick sends a "closed" signal but the channel is already unassigned,
we just log that and do nothing.