Add a power transmission channel to the ant+ classes.
Using this channel, we can transmit power like a power meter, which can be
picked up and recorded by head units (e.g. Garmin Edge 800). This opens
the opportunity for using virtual power as a power meter.

The code has not been added to Big Ring yet, only to the ANT+ test application.
We already have a Sensors class, perhaps we should have an "Actuators"
or "Transmitters" class for sending information to the ANT+ network.