Only emit cadence values from ANT+ Power channels when cadence is valid.
from the power device profile:

9.3 Instantaneous Cadence

The instantaneous cadence field is used to report the pedaling cadence recorded from the power sensor. This is an
instantaneous value only and does not accumulate between messages. The value 0xFF is sent in this field to indicate that
the power sensor cannot measure pedaling cadence. 0xFF is interpreted as an invalid value and is ignored by the display.

This fixes issue #41