We can now use a cache to save and load rlv.
This makes loading gpx and xml files a lot faster. The first time, a file
is always loaded using the normal parser, but it is saved to a binary
cache file after that. If the rlv can be read from this cache file, it
is quickly done, instead of the slow parsing from xml or gpx.

This is a partial fix for #40.