Check version of application when loading a cache file.
This prevents accidentally trying to load a file that does not have the same
format anymore. The only drawback is that when a new version is run, all
cache files will have to be rebuilt. This will only be a problem if we
ever do continuous updates, which is somewhere in the future.